requirements.txt
ctypes
libclang
num2words
